LibWeb: Move AgentType enum to its own header
